fix: Sonar feedback
