+++ kDriveCore/Utils/Logging.swift
@@ -34,7 +34,7 @@ public enum Logging {
         initNetworkLogging()
         initSentry()
         copyDebugInformations()
-        
+
         Log.appDelegate("initLogging done")
     }
